Complete initial draft of the tool
We now have specific :class:`SCons.Builder.Builder`s for SVG to PNG,
PDF, and PDF + LaTeX along with PDF to SVG.  The build parameters are
specified in the ``_params`` dictionary and are dynamically built.
Adding another builder (say SVG to EPS) is a matter of adding the
parameters to the dictionary.